requirements.txt
Django==1.8.13
django-filter==1.1.0
djangorestframework==3.3.3
pkg-resources==0.0.0
Pygments==2.3.1
